fix: Resolve all remaining critical issues from current-issues.txt
- Fix RNG Predictability by rotating ServerSeed every round (Issue 16)
- Fix QuestService transactions by ensuring DB update precedes notifications (Issue 19)
- Fix Sync-over-Async in EfGameRepository by removing blocking Redis calls (Issue 28)
- Fix EF Core Dirty State by detaching entities before update (Issue 29)
- Fix Admin Panel performance by removing heavy GroupBy aggregation (Issue 38)
- Fix VaultController ClaimCashback to be fully asynchronous (Issue 28)
- Clean up EfGameRepository class structure and braces
- Verified fixes for 40+ other issues including Referral collisions, JWT security, and State Recovery

+++ AleaSim.Api/Controllers/AdminController.cs
@@ -38,8 +38,8 @@ public class AdminController : ControllerBase {
         var result = trend.Select(t => new RtpTrendPoint {
             Label = t.Hour.ToString("HH:mm"),
             Rtp = t.Bets > 0 ? (double)(t.Wins / t.Bets) * 100 : 95.0,
-            Bets = (double)t.Bets,
-            Wins = (double)t.Wins
+            Bets = t.Bets,
+            Wins = t.Wins
         }).ToList();
 
         return Ok(result);
